Refactor card props
