+++ src/components/base-components/BaseRequestComponent.jsx
@@ -3,12 +3,12 @@ import { StyleSheet, View } from 'react-native';
 import { Text } from 'react-native-paper';
 
 
-const BaseRequestComponent = ({ posts, CardComponent, title }) => {
+const BaseRequestComponent = ({ posts, Card, title }) => {
     return (
         <View>
             <Text style={styles.textStyle} variant="titleLarge">{title} ({posts.length})</Text>
             {posts.map((post, index) => (
-                <CardComponent key={index} {...post} />
+                <Card key={index} post={post} />
             ))}
         </View>
     )
