+++ src/components/media-associate/AssignedPostsComponent.jsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 
 import AssignedPostCard from './AssignedPostCard';
-import BaseRequestComponent from '../BaseRequestComponent';
+import BaseRequestComponent from '../base-components/BaseRequestComponent';
 
 
 const AssignedPostsComponent = ({ posts }) => {
     return (
         <BaseRequestComponent
             posts={posts}
-            CardComponent={AssignedPostCard}
+            Card={AssignedPostCard}
             title="Current assignments"
         />
     )
