+++ src/components/pending-approval-components/PendingApprovalComponent.jsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 
 import PendingPostCard from './PendingPostCard';
-import BaseRequestComponent from './BaseRequestComponent';
+import BaseRequestComponent from '../base-components/BaseRequestComponent';
 
 
 const PendingApprovalComponent = ({ posts }) => {
     return (
         <BaseRequestComponent
             posts={posts}
-            CardComponent={PendingPostCard}
+            Card={PendingPostCard}
             title="Pending approvals"
         />
     )
