+++ src/components/pending-approval-components/PendingPostCard.jsx
@@ -3,17 +3,17 @@ import * as React from 'react';
 import BaseCard from '../base-components/BaseRequestCard';
 
 
-const AssignedPostCard = ({ post }) => {
+const PendingPostCard = ({ post }) => {
 
     return (
         <BaseCard
             post={post}
-            showDepartment={true}
+            showDepartment={false}
             showReceivedTime={true}
             showPostTime={true}
-            showAssignedTo={false}
+            showAssignedTo={true}
         />
     )
 };
 
-export default AssignedPostCard;
+export default PendingPostCard;
